long long to int_64t

+++ PotreeConverter/include/BoostBINPointReader.hpp
@@ -93,7 +93,7 @@ public:
 
 	AABB getAABB();
 
-	long long numPoints();
+	int64_t numPoints();
 
 	void close();
